requirements.txt
discord.py==2.3.1
icalendar==5.0.7
pylev==1.4.0
Requests==2.31.0
